drm/xe: Allow scanout of bos that can be migrated to lmem
OpenGL stack makes scanout buffers exported as well because usually
it will be exported from client application to compositor.

Exported buffers needs to be placed in SMEM so it can be accessed
by different GPUs if needed, so in discrete GPUs it is placed in
SMEM+LMEM.

But this combination is causing aplications like kmscube to not be
able to create framebuffers.

So here doing the same handling as i915 and allowing bos that can
migrate to lmem0/vram0 to be promoted to framebuffer.

At least all current discrete GPUs with display only have one lmem
region, so we don't need to check for vram1 but we might need to
extend it in future.

Reviewed-by: Rodrigo Vivi <rodrigo.vivi@intel.com>
Signed-off-by: José Roberto de Souza <jose.souza@intel.com>
Signed-off-by: Matthew Auld <matthew.auld@intel.com>