fixup! drm/xe: Introduce a new DRM driver for Intel GPUs
drm/xe: Fix extobj dropping issue.

I believe originally we did the right thing, but now we hit an issue.

Presumably, this was changed when converting from an array to a list
in a squashed commit:
"drm/xe: Use a list for the vm's external objects"

When deleting a vma with an extobj in the list, ensure if it's freed
but still listed on the vm, we take a different vma and add that to
the extobj list instead.

I'll add an igt, and make sure it doesn't happen again.

Signed-off-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>
Closes: https://gitlab.freedesktop.org/drm/xe/kernel/-/issues/194
Reviewed-by: Thomas Hellström <thomas.hellstrom@linux.intel.com>